ik mp analysis

+++ src/scripts/developing/pddlstream/test_arg_utils.py
@@ -14,7 +14,7 @@ def parse_test_args():
     parser.add_argument('--data_idx', type=int, help='data file index')
     parser.add_argument('--cname', type=str, help='checker type', default="None")
 
-    parser.add_argument('--TIMEOUT_MOTION', type=int, default=5, help='motion planning timeout')
+    parser.add_argument('--TIMEOUT_MOTION', type=float, default=5, help='motion planning timeout')
     parser.add_argument('--IK_TRY_NUM', type=int, default=10, help='max. number of trials for ik')
     parser.add_argument('--IK_TIMEOUT_SINGLE', type=float, default=0.01, help='timeout for single ik trial')
     parser.add_argument('--MAX_TIME', type=int, default=100, help='TAMP timeout')
